requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysqlclient==2.2.6
PyJWT==2.10.1
redis==5.2.1
Werkzeug==3.1.3
